qreal fix to make kolf build on ARM (note that I also added qreal for the first qMax expression to make sure it does not break if the implict qrealism goes away for one reason or another)
svn path=/trunk/KDE/kdegames/kolf/; revision=1209228

+++ shape.cpp
@@ -156,8 +156,8 @@ b2Shape* Kolf::EllipseShape::createShape()
 	const b2Vec2 c = toB2Vec2(m_rect.center() * Kolf::Box2DScaleFactor);
 	//ensure some minimum size because b2PolygonShape gets confused when its
 	//area is smaller than FLT_EPSILON (TODO: handle rx*ry == 0 differently?)
-	const qreal rx = qMax(m_rect.width() * Kolf::Box2DScaleFactor / 2, 1e-5);
-	const qreal ry = qMax(m_rect.height() * Kolf::Box2DScaleFactor / 2, 1e-5);
+	const qreal rx = qMax(qreal(m_rect.width() * Kolf::Box2DScaleFactor / 2), qreal(1e-5));
+	const qreal ry = qMax(qreal(m_rect.height() * Kolf::Box2DScaleFactor / 2), qreal(1e-5));
 	if (rx == ry)
 	{
 		//use circle shape when possible because it's cheaper and exact
